added sizes as boxes and font sizes
