+++ src/app/Components/ComponentsSamples/Loading/LoadingSamples.xaml
@@ -7,13 +7,12 @@
              xmlns:dui="http://dips.com/mobile.ui"
              xmlns:localizedStrings="clr-namespace:Components.Resources.LocalizedStrings"
              xmlns:loading="clr-namespace:Components.ComponentsSamples.Loading"
+             xmlns:helpers="clr-namespace:Components.Helpers"
+             xmlns:skeleton="clr-namespace:Components.ComponentsSamples.Loading.Skeleton"
              x:Class="Components.ComponentsSamples.Loading.LoadingSamples">
-    <dui:ContentPage.BindingContext>
-        <loading:LoadingSamplesPageViewModel />
-    </dui:ContentPage.BindingContext>
     <dui:VerticalStackLayout>
         <dui:NavigationListItem Title="{x:Static localizedStrings:LocalizedStrings.SkeletonLoading}"
-                                Command="{Binding NavigateToSkeletonLoadingCommand}"
+                                Command="{helpers:NavigationCommand {x:Type skeleton:SkeletonLoadingSamplesPage}}"
                                 />
     </dui:VerticalStackLayout>
 </dui:ContentPage>
