--- src/app/Components/ComponentsSamples/Loading/LoadingSamplesPageViewModel.cs
@@ -1,13 +1,0 @@
-using System.Windows.Input;
-using Components.ComponentsSamples.Loading.Skeleton;
-
-namespace Components.ComponentsSamples.Loading;
-
-public class LoadingSamplesPageViewModel
-{
-    public LoadingSamplesPageViewModel()
-    {
-        NavigateToSkeletonLoadingCommand = new Command((() => Shell.Current.Navigation.PushAsync(new SkeletonLoadingSamplesPage())));
-    }
-    public ICommand NavigateToSkeletonLoadingCommand { get; } 
-}
