+++ src/app/Components/ComponentsSamples/Pickers/PickersSample.xaml
@@ -6,15 +6,12 @@
                  xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
                  xmlns:dui="http://dips.com/mobile.ui"
                  xmlns:pickers="clr-namespace:Components.ComponentsSamples.Pickers"
+                 xmlns:helpers="clr-namespace:Components.Helpers"
                  x:Class="Components.ComponentsSamples.Pickers.PickersSample">
-    <dui:ContentPage.BindingContext>
-        <pickers:PickerSamplesViewModel />
-    </dui:ContentPage.BindingContext>
-
     <dui:VerticalStackLayout>
         <dui:NavigationListItem Title="Item pickers"
-                                Command="{Binding NavigateToItemPickersSamplesCommand}" />
+                                Command="{helpers:NavigationCommand {x:Type pickers:ItemPickersSamples}}" />
         <dui:NavigationListItem Title="Date / time pickers"
-                                Command="{Binding NavigateToDateTimePickersSamplesCommand}" />
+                                Command="{helpers:NavigationCommand {x:Type pickers:DateTimePickersSamples}}" />
     </dui:VerticalStackLayout>
 </dui:ContentPage>
