+++ src/app/Components/ResourcesSamples/Sizes/SizesAsVisualBoxes.xaml
@@ -1,0 +1,42 @@
+<?xml
+    version="1.0"
+    encoding="utf-8"?>
+
+<dui:ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+                 xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+                 xmlns:dui="http://dips.com/mobile.ui"
+                 x:Class="Components.ResourcesSamples.Sizes.SizesAsVisualBoxes"
+                 Title="Sizes as visual elements">
+    <Grid RowDefinitions="Auto, *">
+        <dui:SearchBar Grid.Row="0"
+                       TextChanged="InputView_OnTextChanged" />
+        <dui:CollectionView Grid.Row="1"
+                            Margin="{dui:Sizes size_3}"
+                            ItemsSource="{Binding Source={RelativeSource AncestorType={x:Type dui:ContentPage}}, Path=Sizes}">
+            <dui:CollectionView.ItemTemplate>
+                <DataTemplate>
+                    <Border>
+                        <Grid ColumnSpacing="5"
+                              Margin="{dui:Sizes size_2}"
+                              ColumnDefinitions="75,50, Auto">
+                            <dui:Label Grid.Column="0"
+                                       VerticalTextAlignment="Center"
+                                       HorizontalOptions="Start"
+                                       Text="{Binding Key}" />
+                            <dui:Label Grid.Column="1"
+                                       VerticalTextAlignment="Center"
+                                       HorizontalOptions="Start"
+                                       Text="{Binding Value}" />
+                            <Border Grid.Column="2"
+                                    WidthRequest="{Binding Value}"
+                                    HorizontalOptions="Start"
+                                    IsVisible="{Binding Value, Converter={dui:IsEmptyConverter Inverted=True}}"
+                                    HeightRequest="{dui:Sizes size_5}"
+                                    BackgroundColor="{dui:Colors color_primary_90}" />
+                        </Grid>
+                    </Border>
+                </DataTemplate>
+            </dui:CollectionView.ItemTemplate>
+        </dui:CollectionView>
+    </Grid>
+</dui:ContentPage>
